Compilation fixes
Fixes to get successful compile on Linux Mint 20.1

+++ data_source_stdio_info.cpp
@@ -3,5 +3,5 @@
 
 void data_source_stdio_info::write( const uint8_t * data, size_t bytes )
 {
-fprintf(stdout,"packet %i: %i bytes\n", ++num_packets, bytes);
+  fprintf(stdout,"packet %i: %i bytes\n", ++num_packets, (int) bytes);
 }
